Callback methods stay the same when using annotations
When using annotations, the Validate and Invalidate annotations stay at
their place, whereas the associated method name changes, which is
incompatible with the filtering technique.
By now, the meta data are modified to use the original method name in
the callback element.
Some equivalent work may be needed on other kinds of elements.