Manifest is updated after the metadata file has been modified
The builder has been re-activated in plugin.xml.
It updates the manifest information if and only if a metadata file has
been modified.
Full builds are now ignored by the builder, as the JDT participant will
do the update in that case.

This may imply two builds, one for the builder (metadata.xml modified)
and one for the JDT (Manifest.mf modified).